Change SWAP Table

changes the value of a variable in the given row of a SWAP table.

Usage

change_swap_table(table, variable, row, value, verbose = F)

Arguments

table

can be one SWAP table, or a list of SWAP tables as loaded by load_swap_tables().

variable

the name of the variable you would like to alter (string)

row

the row you would like to alter (integer)

value

the value you would like to enter (must be SWAP-FORTRAN compatible)

verbose

print status? (flag)

Value

Returns the same table / list of tables passed to the function, with the desired variable modified at the given row.

See also