Estimate Actual Vapour Pressure

Converts relative humidity (RH) to actual vapour pressure (AVP) using an estimation with maximum and minimum temperature. This function exists because SWAP uses AVP whereas many datasets contain RH. Read more on fao.org

Usage

est_avp(tmin, tmax, rh)

Arguments

tmin

numeric vector, in Celsius (C)

tmax

numeric vector, in Celsius (C)

rh

numeric vector, in % (ie. 53.2)

Value

vector of AVP, rounded to 2 decimal placews

Details

Note: if you would like to use this for hourly AVP, then just use the same value for tmin and tmax

See also

Examples

est_avp(tmin = c(1.8, 1.9, 2.0), tmax =  c(6, 7, 8),
                   rh = c(60.2, 40.5, 90.2))
#> [1] 0.49 0.34 0.80