Get Model Performance

Returns the performance for the chosen variable(s) and performance indicator(s). performance indicators are sourced form package hydroGOF Supported variables are those provided in the observed data file.

Usage

get_swap_performance(
  project_path,
  stat,
  archived = FALSE,
  variable = NULL,
  depth = NULL,
  verbose = F
)

Arguments

project_path

path to project directory (string)

stat

statistical performance indicator from package hydroGOF (string)

archived

set to TRUE if project is saved in "rswap_saved/" (flag)

variable

variable name. leave blank for all. (string)

depth

depth of variable if it is depth-wise. leave blank for all (numeric)

verbose

print status? (flag)

Value

dataframe value(s) of performance indicator(s) for given variable(s) and depth(s)

Details

The performance indicator will be the result of a function passed as a string in the stat parameter as called from the hydroGOF package.

The following are supported for stat:

br2 cp d KGE mae md me mNSE mse nrmse NSE pbias rd rmse rNSE rPearson rSD rsr ssq VE

Please note, passing multiple variables at differing depths will not work! (yet)