Load Observed Data

This function loads your observed data from the template file that was placed in your project directory either by build_rswap_directory() or rswap_init()

load_observed(project_path, archived = F, verbose = F)

Arguments

project_path

Path to project directory (string)

archived

set to true if project path in saved in 'rswap_saved' (flag)

verbose

print status? (flag)

Value

Returns a list consiting of .$data, a dataframe of the observed values, as well as .$observed_variables, a vector of the detected observed variables.

Details

It is critical that the template observed file is filled out correctly. Please see the file itself for more information. It should be located in your project directory, and must bear the namer swap_observed_data.csv.

Please note, the file type will eventually be switched to .csv.