Runs the SWAP model

Runs the SWAP model

Usage

run_swap(
  project_path,
  swap_file = "swap.swp",
  autoset_output = F,
  verbose = F,
  timeout = Inf,
  quiet = F
)

Arguments

project_path

String, path to the project directory.

autoset_output

If set to true, rswap will automatically detect your observed data provided in the observed file and match it to the SWAP output. if this is set to false, then INLIST csv must be set by the user either manually or with set_swap_output() or change_swap_par()

verbose

logical, print status?

timeout

number of seconds before run timeout (optional, unlimited by default)

quiet

logical, turn off warnings?

string

name of the *.swp main file

Value

returns name of run (change this!)